+++ apps/core/src/app/view/components/dialogs/DeviceInfoDialog.tsx
@@ -14,7 +14,6 @@ const DeviceInfoDialog = ({ device, onClose, hole }: DeviceInfoDialogProps) => {
 
   const badges = (
     <>
-      <span className="px-[0.625rem] py-1 rounded-full bg-indigo-500/20 text-indigo-200 border border-indigo-400/30 backdrop-blur-sm">ID: {device.id}</span>
       {device.deviceCategory?.name && (
         <span className="px-[0.625rem] py-1 rounded-full bg-emerald-500/20 text-emerald-200 border border-emerald-400/30 backdrop-blur-sm">
           {device.deviceCategory.name}
@@ -28,16 +27,8 @@ const DeviceInfoDialog = ({ device, onClose, hole }: DeviceInfoDialogProps) => {
       <div className="grid grid-cols-2 gap-4 text-sm text-slate-200">
         <div className="space-y-3">
           <div>
-            <span className="text-slate-400 font-medium">장치명</span>
-            <p className="text-slate-100 mt-1">{device.name || '알 수 없음'}</p>
-          </div>
-          <div>
             <span className="text-slate-400 font-medium">장치 ID</span>
             <p className="text-slate-100 mt-1 font-mono">{device.id}</p>
-          </div>
-          <div>
-            <span className="text-slate-400 font-medium">카테고리</span>
-            <p className="text-slate-100 mt-1">{device.deviceCategory?.name || '미지정'}</p>
           </div>
         </div>
       </div>
