+++ apps/core/src/app/view/components/dialogs/FacilityInfoDialog.tsx
