+++ apps/core/src/app/view/components/dialogs/index.ts
@@ -1,0 +1,3 @@
+export { default as FacilityInfoDialog } from './FacilityInfoDialog'
+export { default as DeviceInfoDialog } from './DeviceInfoDialog'
+export { default as CctvDialog } from './CctvDialog'
