--- apps/core/src/app/view/components/info-dialog/index.ts
@@ -1,2 +1,0 @@
-export { default as FacilityInfoDialog } from './FacilityInfoDialog'
-export { default as DeviceInfoDialog } from './DeviceInfoDialog'
