+++ apps/core/src/app/view/components/maps/OutdoorMap.tsx
@@ -5,7 +5,7 @@ import type { FacilityType, FacilityResponse } from '@plug/common-services'
 import { VWorldMap, MapControls, CameraSetup } from '@/global/components/outdoor-map'
 import FacilityPOIs from './FacilityPOIs'
 import FacilitySearchForm from './FacilitySearchForm'
-import FacilityInfoDialog from '../info-dialog/FacilityInfoDialog'
+import FacilityInfoDialog from '../dialogs/FacilityInfoDialog'
 
 interface OutdoorMapProps {
   onFacilitySelect?: (facilityId: number, facilityType: FacilityType) => void;
